+++ structures/models/notifications.py
@@ -1,6 +1,5 @@
 """Notification related models"""
 
-from datetime import timedelta
 import logging
 import yaml
 from typing import Tuple
@@ -76,7 +75,7 @@ LANGUAGES = (
 
 
 class NotificationType(models.TextChoices):
-    """Definition of all supported notification types"""
+    """Definition of all supported notification types."""
 
     # upwell structures
     STRUCTURE_ANCHORING = "StructureAnchoring", _("Upwell structure anchoring")
@@ -170,12 +169,20 @@ class NotificationType(models.TextChoices):
     )
 
     # corporation membership
+    CORP_APP_NEW_MSG = "CorpAppNewMsg", _("Character submitted application")
+    CORP_APP_INVITED_MSG = "CorpAppInvitedMsg", _(
+        "Character invited to join corporation"
+    )
+    CORP_APP_REJECT_CUSTOM_MSG = "CorpAppRejectCustomMsg", _(
+        "Corp application rejected"
+    )
+    CHAR_APP_WITHDRAW_MSG = "CharAppWithdrawMsg", _("Character withdrew application")
     CHAR_APP_ACCEPT_MSG = "CharAppAcceptMsg", _("Character joins corporation")
     CHAR_LEFT_CORP_MSG = "CharLeftCorpMsg", _("Character leaves corporation")
 
     @classproperty
     def webhook_defaults(cls) -> list:
-        """list of default notifications for new webhooks"""
+        """List of default notifications for new webhooks."""
         return [
             cls.STRUCTURE_ANCHORING,
             cls.STRUCTURE_DESTROYED,
@@ -200,7 +207,6 @@ class NotificationType(models.TextChoices):
         return [
             cls.STRUCTURE_LOST_SHIELD,
             cls.STRUCTURE_LOST_ARMOR,
-            cls.STRUCTURE_ANCHORING,
             cls.ORBITAL_REINFORCED,
             cls.MOONMINING_EXTRACTION_STARTED,
             cls.MOONMINING_EXTRACTION_CANCELLED,
@@ -251,12 +257,12 @@ class NotificationType(models.TextChoices):
 
 
 def get_default_notification_types():
-    """DEPRECATED: generates a set of all existing notification types as default"""
+    """DEPRECATED: generates a set of all existing notification types as default-"""
     return tuple(sorted([str(x[0]) for x in NotificationType.values]))
 
 
 class Webhook(WebhookBase):
-    """A destination for forwarding notification alerts"""
+    """A destination for forwarding notification alerts."""
 
     notification_types = MultiSelectField(
         choices=NotificationType.choices,
@@ -296,7 +302,7 @@ class Webhook(WebhookBase):
 
 
 class EveEntity(models.Model):
-    """An EVE entity like a character or an alliance"""
+    """An EVE entity like a character or an alliance."""
 
     CATEGORY_CHARACTER = 1
     CATEGORY_CORPORATION = 2
@@ -327,7 +333,7 @@ class EveEntity(models.Model):
         )
 
     def profile_url(self) -> str:
-        """returns link to website with profile info about this entity"""
+        """Returns link to website with profile info about this entity."""
         if self.category == self.CATEGORY_CORPORATION:
             url = dotlan.corporation_url(self.name)
         elif self.category == self.CATEGORY_ALLIANCE:
@@ -351,7 +357,7 @@ class EveEntity(models.Model):
 
     @classmethod
     def get_matching_entity_category(cls, type_name) -> int:
-        """returns category for given ESI name"""
+        """Returns category for given ESI name."""
         match = None
         for x in cls.CATEGORY_CHOICES:
             if type_name == x[1]:
@@ -361,7 +367,7 @@ class EveEntity(models.Model):
 
 
 class Notification(models.Model):
-    """An EVE Online notification about structures"""
+    """An EVE Online notification about structures."""
 
     HTTP_CODE_TOO_MANY_REQUESTS = 429
 
@@ -431,17 +437,17 @@ class Notification(models.Model):
 
     @property
     def is_alliance_level(self) -> bool:
-        """whether this is an alliance level notification"""
+        """Whether this is an alliance level notification."""
         return self.notif_type in NotificationType.relevant_for_alliance_level
 
     @property
     def can_be_rendered(self) -> bool:
-        """whether this notification can be rendered in Discord"""
+        """Whether this notification can be rendered in Discord."""
         return self.notif_type in NotificationType.values
 
     @property
     def can_have_timer(self) -> bool:
-        """whether this notification can have a timer"""
+        """Whether this notification can have a timer."""
         return self.notif_type in NotificationType.relevant_for_timerboard
 
     # @classmethod
@@ -450,11 +456,11 @@ class Notification(models.Model):
     #     return {x[0] for x in NotificationType.choices}
 
     def get_parsed_text(self) -> dict:
-        """returns the notifications's text as dict"""
+        """Returns the notifications's text as dict."""
         return yaml.safe_load(self.text)
 
     def is_npc_attacking(self) -> bool:
-        """whether this notification is about a NPC attacking"""
+        """Whether this notification is about a NPC attacking."""
         result = False
         if self.notif_type in [
             NotificationType.ORBITAL_ATTACKED,
@@ -479,15 +485,16 @@ class Notification(models.Model):
         return result
 
     def filter_for_npc_attacks(self) -> bool:
-        """true when notification to be filtered out due to npc attacks"""
+        """True when notification to be filtered out due to npc attacks."""
         return not STRUCTURES_REPORT_NPC_ATTACKS and self.is_npc_attacking()
 
     def filter_for_alliance_level(self) -> bool:
-        """true when notification to be filtered out due to alliance level"""
+        """True when notification to be filtered out due to alliance level."""
         return self.is_alliance_level and not self.owner.is_alliance_main
 
     def send_to_webhook(self, webhook: Webhook) -> bool:
-        """sends this notification to the configured webhook
+        """Sends this notification to the configured webhook.
+
         returns True if successful, else False
         """
         logger.info("%s: Trying to sent to webhook: %s", self, webhook)
@@ -556,7 +563,7 @@ class Notification(models.Model):
     def _generate_embed(
         self, language_code: str
     ) -> Tuple[dhooks_lite.Embed, Webhook.PingType]:
-        """generates a Discord embed for this notification"""
+        """Generates a Discord embed for this notification."""
         from ..core.notification_embeds import NotificationBaseEmbed
 
         logger.info("Creating embed with language = %s" % language_code)
@@ -572,8 +579,9 @@ class Notification(models.Model):
             return None
 
     def process_for_timerboard(self, token: Token = None) -> bool:
-        """add/removes a timer related to this notification for some types
-        returns True when a timer was processed, else False
+        """Add/removes a timer related to this notification for some types
+
+        Returns True when a timer was processed, else False
         """
         timer_created = False
         if (
@@ -589,8 +597,6 @@ class Notification(models.Model):
                         timer_created = self._gen_timer_structure_reinforcement(
                             parsed_text, token
                         )
-                    elif self.notif_type == NotificationType.STRUCTURE_ANCHORING:
-                        timer_created = self._gen_timer_structure_anchoring(parsed_text)
                     elif self.notif_type == NotificationType.SOV_STRUCTURE_REINFORCED:
                         timer_created = self._gen_timer_sov_reinforcements(parsed_text)
                     elif self.notif_type == NotificationType.ORBITAL_REINFORCED:
@@ -632,7 +638,7 @@ class Notification(models.Model):
     def _gen_timer_structure_reinforcement(
         self, parsed_text: str, token: Token
     ) -> bool:
-        """generate timer for structure reinforcements"""
+        """Generate timer for structure reinforcements"""
         structure_obj, _ = Structure.objects.get_or_create_esi(
             parsed_text["structureID"], token
         )
@@ -688,60 +694,8 @@ class Notification(models.Model):
 
         return timer_added
 
-    def _gen_timer_structure_anchoring(self, parsed_text: str) -> bool:
-        """generate timer for structure anchoring"""
-        structure_type, _ = EveType.objects.get_or_create_esi(
-            parsed_text["structureTypeID"]
-        )
-        solar_system, _ = EveSolarSystem.objects.get_or_create_esi(
-            parsed_text["solarsystemID"]
-        )
-        timer_added = False
-        if not solar_system.is_null_sec:
-            eve_time = self.timestamp + timedelta(hours=24)
-            if has_auth_timers:
-                AuthTimer.objects.create(
-                    details=gettext("Anchor timer"),
-                    system=solar_system.name,
-                    planet_moon="",
-                    structure=structure_type.name,
-                    objective="Friendly",
-                    eve_time=eve_time,
-                    eve_corp=self.owner.corporation,
-                    corp_timer=STRUCTURES_TIMERS_ARE_CORP_RESTRICTED,
-                )
-                timer_added = True
-
-            if has_structure_timers:
-                eve_solar_system, _ = EveSolarSystem2.objects.get_or_create_esi(
-                    id=parsed_text["solarsystemID"]
-                )
-                structure_type, _ = EveType2.objects.get_or_create_esi(
-                    id=parsed_text["structureTypeID"]
-                )
-                visibility = (
-                    Timer.VISIBILITY_CORPORATION
-                    if STRUCTURES_TIMERS_ARE_CORP_RESTRICTED
-                    else Timer.VISIBILITY_UNRESTRICTED
-                )
-                Timer.objects.create(
-                    eve_solar_system=eve_solar_system,
-                    structure_type=structure_type,
-                    timer_type=Timer.TYPE_ANCHORING,
-                    objective=Timer.OBJECTIVE_FRIENDLY,
-                    date=eve_time,
-                    eve_corporation=self.owner.corporation,
-                    eve_alliance=self.owner.corporation.alliance,
-                    visibility=visibility,
-                    owner_name=self.owner.corporation.corporation_name,
-                    details_notes=self._timer_details_notes(),
-                )
-                timer_added = True
-
-        return timer_added
-
     def _gen_timer_sov_reinforcements(self, parsed_text: str) -> bool:
-        """generate timer for sov reinforcements"""
+        """Generate timer for sov reinforcements."""
         if not self.owner.is_alliance_main:
             return False
 
@@ -800,7 +754,7 @@ class Notification(models.Model):
         return timer_added
 
     def _gen_timer_orbital_reinforcements(self, parsed_text: str) -> bool:
-        """generate timer for orbital reinforcements"""
+        """Generate timer for orbital reinforcements."""
         solar_system, _ = EveSolarSystem.objects.get_or_create_esi(
             parsed_text["solarSystemID"]
         )
@@ -851,7 +805,7 @@ class Notification(models.Model):
         return timer_added
 
     def _gen_timer_moon_extraction(self, parsed_text: str) -> bool:
-        """generate timer for moon mining extractions"""
+        """Generate timer for moon mining extractions."""
         solar_system, _ = EveSolarSystem.objects.get_or_create_esi(
             parsed_text["solarSystemID"]
         )
@@ -976,7 +930,7 @@ class Notification(models.Model):
         return timer_added
 
     def _timer_details_notes(self) -> str:
-        """returns generated details notes string for Timers"""
+        """Return generated details notes string for Timers."""
         return (
             "Automatically created from structure notification for "
             f"{self.owner.corporation} at {self.timestamp.strftime(DATETIME_FORMAT)}"